moved events to dedicated directory Fixed transaction > Transaction Entity: missing @Column annotations
